DB Stuff added
If you don't have a local Postgres installation, just go to the
application.conf file and set db=mem. It'll create a new database each
time you run the app.

I also had to change User to UserAccount, for some reason User would not
work. I think it might've been due to a table of that name already
existing in Postgres, and I'm not used to managing schemas in Postgres,
so I figured it wouldn't hurt to just change the name.

I also created a very very simple view called success.html to show how
the intended workflow is for Play after you handle a POST. You can see
this in the Routes file.

I turned Tag into its own class so that we can inherit Model and use all
the Hibernation stuff on the tags as well.

I created a User create form in index.html that exemplifies how to
create users. We can move, modify or do whatever with that.

I was going to do the same for Video but I don't have any more time for
that today.