Runtime fixes
I modified this so it would run on my computer. I renamed the id field
in the video model because it was causing a DB conflict with the Id
column that is automatically created by Play/Hibernate, now it's called
vidId.

Additionally I added jquery 1.7.1 to the javascripts file and changed
the entry in main.html to reflect that change. If you're running an
older jquery you'll have to change that for the app to run in your
browser.